Move userGuidePath helper into content/types.ts
Per review feedback, a standalone paths.ts for one small helper felt
like over-fragmentation. Co-locate userGuidePath with the GroupSlug/
PageSlug values it builds a route from (types.ts is a dependency-free
leaf module, so no circular imports). Remove paths.ts.
